+++ package.json
@@ -11,6 +11,8 @@
         "vite": "^5"
     },
     "dependencies": {
+        "@types/react": "^18.3.1",
+        "@types/react-dom": "^18.3.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1"
     }
