+++ resources/js/@types/global.d.ts
@@ -1,0 +1,9 @@
+import { Data } from "type";
+
+export { };
+
+declare global {
+    interface Window {
+        data: Data
+    }
+}
