+++ resources/views/index.blade.php
@@ -6,5 +6,5 @@
 window.data = @json($data);
 </script>
 @viteReactRefresh
-@vite('resources/js/app.jsx')
+@vite('resources/js/app.tsx')
 @endsection
