+++ vite.config.js
@@ -5,7 +5,7 @@ import react from '@vitejs/plugin-react';
 export default defineConfig({
     plugins: [
         laravel({
-            input: ['resources/js/app.jsx'],
+            input: ['resources/js/app.tsx'],
             refresh: true,
         }),
         react(),
